requirements.txt
# Core
aiohttp>=3.9.0
aiosqlite>=0.19.0
redis>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM APIs (using REST, no SDK needed)
httpx>=0.25.0

# Detection
scikit-learn>=1.3.0
numpy>=1.24.0

# Communication
aiosmtplib>=3.0.0

# Utilities
structlog>=23.2.0
tenacity>=8.2.0
colorama>=0.4.6

# ML requirements
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
structlog>=23.0.0

# gRPC requirements
grpcio==1.60.0
grpcio-tools==1.60.0
docker==7.0.0

streamlit>=1.30.0